Move tests to ConnectionManagementCommandTests
